refactor: 모델에 participant 패키지 생성, 파일 이동
