+++ src/main/kotlin/blackjack/BlackjackGame.kt
@@ -1,7 +1,7 @@
 package blackjack
 
 import blackjack.controller.BlackjackController
-import blackjack.model.Dealer
+import blackjack.model.participant.Dealer
 import blackjack.view.InputView
 import blackjack.view.OutputView
