+++ src/test/kotlin/blackjack/DealerTest.kt
@@ -1,7 +1,7 @@
 package blackjack
 
 import blackjack.model.card.Card
-import blackjack.model.Dealer
+import blackjack.model.participant.Dealer
 import blackjack.model.card.CardNumber
 import blackjack.model.card.Shape
 import org.assertj.core.api.Assertions.assertThat
