+++ src/test/kotlin/blackjack/PlayerTest.kt
@@ -2,7 +2,7 @@ package blackjack
 
 import blackjack.model.card.Card
 import blackjack.model.card.CardNumber
-import blackjack.model.Player
+import blackjack.model.participant.Player
 import blackjack.model.card.Shape
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.BeforeEach
